+++ modules/gui/qt/medialibrary/mlvideomodel.cpp
@@ -55,7 +55,7 @@ MLVideoModel::MLVideoModel(QObject* parent)
 
 QVariant MLVideoModel::data(const QModelIndex& index, int role) const
 {
-    const auto video = item(static_cast<unsigned int>(index.row()));
+    const auto video = item(index.row());
     if ( video == nullptr )
         return {};
     switch (role)
